\name{available_authorities}
\alias{available_authorities}
\title{Returns a character vector corresponding to the currently available authorities.}
\value{
  A list containing the currently available authorities.
}
\description{
  Returns a character vector corresponding to the currently
  available authorities.
}
\examples{
\dontrun{
library('SynergizerR')
available_authorities() }
}
\references{
  http://llama.mshri.on.ca/synergizer/translate/
}

