\name{server_version}
\alias{server_version}
\title{Returns current version of the Synergizer server as a string.}
\value{
  A one-element character vector containing current version
  of the Synergizer server
}
\description{
  Returns current version of the Synergizer server as a
  string.
}
\examples{
\dontrun{
library('SynergizerR')
server_version() }
}
\references{
  http://llama.mshri.on.ca/synergizer/translate/
}

